requirements.txt
# AIOHTTP
aiohttp
# Cryptography
PyCryptodome
